% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_refs.R
\name{apply_refs}
\alias{apply_refs}
\title{Apply Reference Levels to a Dataframe}
\usage{
apply_refs(data, cat_vars = NULL, ref_levels = NULL)
}
\arguments{
\item{data}{A data frame with containing the variables specified in the formula}

\item{cat_vars}{A string vector referring to the outcome variable}

\item{ref_levels}{A string vector referring to the reference levels}
}
\value{
A dataframe with applied reference levels
}
\description{
Automatically converts specified columns within a dataframe to factors with specified reference levels
}
\details{
Placeholder
}
