% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{applyNormalization}
\alias{applyNormalization}
\title{Applies normalization methods to given data set}
\usage{
applyNormalization(
  data,
  groups,
  method = c("TC", "UQ", "median", "TMM", "DESeq", "PoissonSeq", "QN", "RUV")
)
}
\arguments{
\item{data}{Un-normalized count data set of shape p x n, where p is the
number of markers n is the number of samples.}

\item{groups}{Vector of length n that maps the samples to sample groups.}

\item{method}{Vector of normalization methods that are applied to the data.
Available methods are: \code{c("TC", "UQ", "median", "TMM", "DESeq",
"PoissonSeq", "QN", "RUV")}. Select one or multiple methods. By default all
normalization methods will be applied.}
}
\value{
List of normalized read counts for selected normalization methods.
}
\description{
Applies single or multiple normalization methods to the given data.
  Available methods are:
  \itemize{
    \item Total Count (TC)
    \item Upper Quartile (UQ)
    \item Median (median)
    \item Trimmed Median of Means (TMM)
    \item DESeq
    \item PoissonSeq
    \item Quantile Normalization (QN)
    \item Remove Unwanted Variation (RUVg, RUVr, and RUVs)
  }
}
