% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_flowcharts.R
\name{binary_flowchart}
\alias{binary_flowchart}
\title{Modify binary flowchart}
\usage{
binary_flowchart(vals, final_node)
}
\arguments{
\item{placeholder}{ph}
}
\value{
placeholder
}
\description{
Highlights flowchart path and inputs test values
}
\details{
Placeholder
}
