% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DANA.R
\name{compute.cc}
\alias{compute.cc}
\title{Compute the concordance correlation of clustered positive controls}
\usage{
compute.cc(rawCor, normCor, clusters)
}
\description{
Compute the concordance correlation of clustered positive controls
}
\keyword{internal}
