% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DANA.R
\name{compute.mcr}
\alias{compute.mcr}
\title{Compute the mean correlation reduction in negative controls}
\usage{
compute.mcr(rawCor, normCor)
}
\description{
Compute the mean correlation reduction in negative controls
}
\keyword{internal}
