% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_flowcharts.R
\name{continuous_flowchart}
\alias{continuous_flowchart}
\title{Modify Continuous flowchart}
\usage{
continuous_flowchart(a, b, final_node)
}
\arguments{
\item{placeholder}{ph}
}
\value{
placeholder
}
\description{
Highlights flowchart path and inputs test values
}
\details{
Placeholder
}
