% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_sectional.R
\name{cross_sectional}
\alias{cross_sectional}
\title{Bayesian Inference for Continuous Outcomes}
\usage{
cross_sectional(
  data,
  prior = "improper",
  family = "gaussian",
  offset_var = NULL,
  outcome_var = NULL,
  pred_cat = NULL,
  pred_cont = NULL,
  ref_levels = NULL,
  ci_level = NULL
)
}
\arguments{
\item{data}{A data frame with containing the variables specified in the formula.}

\item{prior}{A string of either "improper", "zellner", or "conjugate"}

\item{family}{family specifying distributional assumption on outcome, either "gaussian", "poisson", or "binomial"}

\item{offset_var}{A string referring to offset variable. Ignored unless family == "poisson"}

\item{outcome_var}{A string referring to the outcome variable}

\item{pred_cat}{A vector of strings referring to categorical predictor variables}

\item{pred_cont}{A vector of strings referring to continuous predictor variables}

\item{ref_levels}{A string vector specifying the reference levels for all categorical variables}

\item{ci_level}{A numeric value specifying the credible interval level}
}
\value{
A list containing the results from the lm_b fit, along with potential diagnostic plots.
}
\description{
Wrapper for lm_b for continuous outcomes, with additional diagnostics and defaults.
}
\details{
Performs bayesian linear models with improper prior. Also generates
basic diagnostic plots depending on the model type. The plot outputs
are of class ggplot2::ggplot, tables are of class DT::dt, and info is of class list.
The following are provided for each model type:

Gaussian - Table, info, interpretations, QQ plot, homoscedasticity plot
Poisson - Table, info, interpretations
Binomial - Table, info, interpretations
}
