% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_sectional.R
\name{cross_sectional_formula}
\alias{cross_sectional_formula}
\title{Generate Formula Objects for Cross Sectional Regression}
\usage{
cross_sectional_formula(
  family,
  pred_cat,
  pred_cont,
  outcome_var,
  offset_var = NULL
)
}
\arguments{
\item{family}{family specifying distributional assumption on outcome}

\item{pred_cat}{A vector of strings referring to categorical predictor variables}

\item{pred_cont}{A vector of strings referring to continuous predictor variables}

\item{outcome_var}{A string referring to the outcome variable}

\item{offset_var}{A string referring to offset variable. Ignored unless family == "poisson"}
}
\value{
A formula object
}
\description{
Generates formulas based on regression type and selected variables
}
\details{
Manipulates provided string objects into a ready-to-use formula object
}
