% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{defineClusters}
\alias{defineClusters}
\title{Polycistronic clustering of miRNAs}
\usage{
defineClusters(genes, chr, pos, threshold = 10000)
}
\arguments{
\item{genes}{List of genes for which the miRNA clustering is computed.}

\item{chr}{Vector that specifies the chromosome for each miRNA in
\code{genes}.}

\item{pos}{Vector of the base-pair position of each miRNA in \code{genes} on
its chromosome, which is specified in \code{chr}. This position could be
e.g. the start, end or averaged ((start+end)/2) base-pair position on the
associated chromosome.}

\item{threshold}{Maximum distance threshold in base-pairs for the difintion
of polycistronic clusters.}
}
\value{
Vector of clusters. Associates each miRNA in \code{genes} to a
  polycistronic cluster.
}
\description{
Defines polycistronic clusters for a given set of miRNAs based on their
location on the genome.
}
