% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_var_type.R
\name{detect_var_type}
\alias{detect_var_type}
\title{Detect Variable Types}
\usage{
detect_var_type(data)
}
\arguments{
\item{data}{tibble input}
}
\description{
Basic logic for detecting (guessing) variable types based on data input
}
\details{
Returns character vectors corresponding to the column
names of each variable type. Uses basic logic to detect reasonable
types
}
