% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cross_sectional_continuous.R
\name{fit_cross_sectional_continuous}
\alias{fit_cross_sectional_continuous}
\title{Fit Continuous Outcome For Cross Sectional Data}
\usage{
fit_cross_sectional_continuous(
  data,
  prior = "improper",
  response_var = NULL,
  pred_cat_vars = NULL,
  pred_cont_vars = NULL,
  pred_cat_vars_ref_levels = NULL,
  ci_level = NULL
)
}
\arguments{
\item{a}{placeholder}
}
\description{
Fit Continuous Outcome For Cross Sectional Data
}
