% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cross_sectional_regression.R
\name{fit_cross_sectional_regression}
\alias{fit_cross_sectional_regression}
\title{Fit Study}
\usage{
fit_cross_sectional_regression(
  data,
  prior = "improper",
  offset_var = NULL,
  response_var = NULL,
  pred_cat_vars = NULL,
  pred_cont_vars = NULL,
  pred_cat_vars_ref_levels = NULL,
  ci_level = NULL
)
}
\arguments{
\item{id}{shiny input/output id}

\item{plotly_list}{a list of plotly objects}
}
\description{
Return information from model fits and other statistical results
}
\details{
Paired with uiCatScrollPlot through the id argument.
This function is placed in an R shiny
server function with a specified id and list of plotly objects.

When rendered in combination with uiCatScrollPlot, displays plotly objects
from the list, which can be scrolled with the associated ActionButton.
}
\seealso{
\code{\link{uiCatScrollPlot}}
}
