% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_sum.R
\name{gen_sum}
\alias{gen_sum}
\title{General Summary Statistics}
\usage{
gen_sum(data, cont_vars, cat_vars, group_var = NULL)
}
\arguments{
\item{placeholder}{test}
}
\value{
placeholder
}
\description{
Computes summary statistics and graphs ready for input into shiny
}
\details{
placeholder
}
