% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm.QN}
\alias{norm.QN}
\title{Quantile Normalization (QN)}
\usage{
norm.QN(raw, filter = FALSE)
}
\arguments{
\item{raw}{Raw read count matrix (rows = genes, cols = samples).}

\item{groups}{Sample groups (subtypes)}
}
\value{
A list with the following elements:
\describe{
  \item{dataNormalized}{Normalized read counts.}
}
}
\description{
Quantile Normalization (QN) for miRNA-Seq data.
}
\examples{
rawCounts <- matrix(0:20, nrow = 7)
normCounts <- norm.QN(rawCounts)
}
