% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCor.R
\name{partialCor}
\alias{partialCor}
\title{Partial correlation estimation}
\usage{
partialCor(X, scale = TRUE)
}
\arguments{
\item{X}{Data matrix. Size (samples x parameters): n x p. Columns should be
named corresponding to marker/miRNA names.}

\item{scale}{Logical if the data is scaled column-wise to mean 0 and var 1.}
}
\value{
Estimated partial correlation matrix. Shape p x p
}
\description{
Partial correlation estimation using  the neighborhood selection method (see
Meinshausen, Buehlmann (2006)) for the estimation of the inverse covariance
matrix (precision matrix)
}
\references{
\href{https://projecteuclid.org/euclid.aos/1152540754}{Neighborhood
Selection}
}
