% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCor.R
\name{pearsonCor}
\alias{pearsonCor}
\title{Pearson correlation estimation}
\usage{
pearsonCor(X)
}
\arguments{
\item{X}{Data matrix. Size (samples x parameters): n x p}
}
\value{
Estimated Pearson correlation matrix. Shape p x p
}
\description{
Estimaetes pearson correlations for a given data matrix.
}
