% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotCor}
\alias{plotCor}
\title{Correlation plot for partial correlations within polycistronic clusters}
\usage{
plotCor(pCor, clusters, title)
}
\arguments{
\item{pCor}{Matrix of partial correlations. Rows and columns must be
named by their corresponding markers}

\item{clusters}{Named Vector of clusters. Associates each marker
to a polycistronic cluster. Usually generated using
the function \code{\link{defineClusters}}.}

\item{title}{\emph{Optional.} Plot title}
}
\value{
\code{ggplot} object of a correlation plot for the given
partial correlations.
}
\description{
This function generates a correlation (heatmap-type) plot for
partial correlations. Polycistronic clusters of the markers are
highlighted in the lower tri-diagonal part of the matrix while
partial correlations are shown in the upper tri-diagonal part.
Partial correlations can be computed using the \code{\link{partialCor}}
function.
}
