% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotCountHist}
\alias{plotCountHist}
\title{Read count histogram}
\usage{
plotCountHist(raw, binwidth = 0.1, tZero, tPoor, tWell, title)
}
\arguments{
\item{raw}{Raw read count matrix (rows = genes, cols = samples).}

\item{binwidth}{\emph{Optional.} The width of the bins as a numeric value.
Default is 0.1.}

\item{tZero}{\emph{Optional.} Lower bound for negative controls.}

\item{tPoor}{\emph{Optional.} Upper bound for negative controls.}

\item{tWell}{\emph{Optional.} Lower bound for positive controls.}

\item{title}{Plot title}

\item{xlim.max}{Upper limit of the x-axis}

\item{ylim.max}{Upper limit of the y-axis}
}
\value{
ggplot2 object of the read count histogram
}
\description{
Plots a histogram of the marker-specific mean of log2-transformed
read counts. Bounds for the definition of positive
and negative controls can be highlighted by vertical lines (if given).
}
\examples{
counts <- matrix(rnbinom(1000, mu=3, size=0.01), nrow=50)
plotCountHist(counts)
}
