% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotMeanSD}
\alias{plotMeanSD}
\title{Mean-Standard Deviation Plot}
\usage{
plotMeanSD(raw, tZero, tPoor, tWell, title, xlim.max = NA, ylim.max = NA)
}
\arguments{
\item{raw}{Raw read count matrix (rows = genes, cols = samples).}

\item{tZero}{\emph{Optional.} Lower bound for negative controls.}

\item{tPoor}{\emph{Optional.} Upper bound for negative controls.}

\item{tWell}{\emph{Optional.} Lower bound for positive controls.}

\item{title}{Plot title}

\item{xlim.max}{Upper limit of the x-axis}

\item{ylim.max}{Upper limit of the y-axis}
}
\value{
ggplot object of the mean-sd plot
}
\description{
Plots the marker-specific standard deviation over marker-specific means
for log2-normalized read counts. Bounds for the definition of positive
and negative controls can be highlighted by vertical lines (if given).
}
\examples{
counts <- matrix(rnbinom(1000, mu=3, size=0.01), nrow=50)
plotMeanSD(counts)
}
