% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateCor.R
\name{prec2corr}
\alias{prec2corr}
\title{Computes partial correlation matrix from a given precision matrix}
\usage{
prec2corr(P)
}
\arguments{
\item{P}{A square precision matrix}
}
\value{
Partial correlation matrix computed from P
}
\description{
Computes partial correlation matrix from a given precision matrix
}
