% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read Data from File}
\usage{
read_data(file, start_row, end_row)
}
\arguments{
\item{start_row}{Number indicating row with variable names}

\item{end_row}{Number indicating row with final observation}

\item{path}{Path pointing to file to read}
}
\value{
Dataframe
}
\description{
Reads data from a chosen fileInput
}
\details{
Designed to read 3 different file formats: csv, xlsx, and rds.
These are specifically chosen to be most common formats in public
health research.

This function has two behaviors:
\itemize{
\item If start_row and end_row are NA, then use defaults to read in
data.
\item If start_row and end_row are specified, then treats the start
row as row labels and reads to the specified data rows.
}
}
