% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrospective_cohort.R
\name{retrospective_cohort_regression}
\alias{retrospective_cohort_regression}
\title{Bayesian Inference for Retrospective Cohort Designs}
\usage{
retrospective_cohort_regression(
  data,
  family = "gaussian",
  outcome_var,
  time_var,
  pred_cat,
  pred_cont,
  group_var,
  ref_levels = NULL,
  offset_var = NULL,
  ci_level = NULL
)
}
\arguments{
\item{data}{A data frame with containing the variables specified in the formula}

\item{family}{family specifying distributional assumption on outcome}

\item{outcome_var}{A string referring to the outcome variable}

\item{time_var}{A string referring to the time variable}

\item{pred_cat}{A vector of strings referring to categorical predictor variables}

\item{pred_cont}{A vector of strings referring to continuous predictor variables}

\item{group_var}{A string referring to the group variable}

\item{ref_levels}{A string vector specifying the reference levels for all categorical variables}

\item{offset_var}{A string referring to offset variable. Ignored unless family == "poisson"}

\item{ci_level}{A numeric value specifying the credible interval level}
}
\value{
A list containing the results from the stan_glm.
}
\description{
Wrapper for stan_glmer for cohort designs, with additional diagnostics and defaults.
}
\details{
Performs bayesian generalized linear mixed-effects models.
}
