% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleCatScrollPlot.R
\name{serverCatScrollPlot}
\alias{serverCatScrollPlot}
\title{CatScrollPlot Server Logic}
\usage{
serverCatScrollPlot(id, data, cat_vars, group_var)
}
\arguments{
\item{id}{shiny input/output id}

\item{plotly_list}{a list of plotly objects}
}
\description{
Create server logic for scrolling plotly outputs based on categorical variables
}
\details{
Paired with uiCatScrollPlot through the id argument.
This function is placed in an R shiny
server function with a specified id and list of plotly objects.

When rendered in combination with uiCatScrollPlot, displays plotly objects
from the list, which can be scrolled with the associated ActionButton.
}
\seealso{
\code{\link{uiCatScrollPlot}}
}
