% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleContScrollPlot.R
\name{serverContScrollPlot}
\alias{serverContScrollPlot}
\title{ContScrollPlot Server Logic}
\usage{
serverContScrollPlot(id, data, cont_vars, group_var = NULL)
}
\arguments{
\item{id}{shiny input/output id}

\item{plotly_list}{a list of plotly objects}
}
\description{
Create server logic for scrolling plotly outputs based on continuous variables
}
\details{
Paired with uiContScrollPlot through the id argument.
This function is placed in an R shiny
server function with a specified id and list of plotly objects.
When rendered in combination with uiContScrollPlot, creates a dynamic
plotly object with actionButtons that scroll through the list.
}
