% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleDataTable.R
\name{serverDataTable}
\alias{serverDataTable}
\title{DataTable Server Logic}
\usage{
serverDataTable(id, data)
}
\arguments{
\item{id}{shiny input/output id}

\item{data}{dataframe}
}
\description{
Create server logic to generate a table showcasing the uploaded data
}
\details{
Paired with uiDataTable through the id argument.
This function is placed in an R shiny
server function with a specified id and data frame/tibble.
When rendered in combination with uiDataTable, creates an output
data table that can be browsed by the user.
}
