% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleFiveNum.R
\name{serverFiveNum}
\alias{serverFiveNum}
\title{Five Number Summary Server Logic}
\usage{
serverFiveNum(id, data, cat_vars, cont_vars, group_var)
}
\arguments{
\item{id}{shiny input/output id}

\item{plotly_list}{a list of plotly objects}
}
\description{
Create server logic for five number summary table
}
\details{
Paired with uiFiveNum through the id argument.
This function is placed in an R shiny
server function with a specified id and dataset.
When rendered in combination with uiFiveNum, creates a summary
table for the variables in the dataset. Summary table
is created with table1 package.

Includes options for displaying summary by grouping
variable.
}
