% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleFlowchart.R
\name{serverFlowchart}
\alias{serverFlowchart}
\title{Flowchart Server Logic}
\usage{
serverFlowchart(id, vals, final_node)
}
\arguments{
\item{id}{shiny input/output id}
}
\description{
Create server logic for displaying modified flowcharts
}
