% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverGLMMInterpret.R
\name{serverGLMMInterpret}
\alias{serverGLMMInterpret}
\title{GLMMInterpret Server Logic}
\usage{
serverGLMMInterpret(id, analysis)
}
\arguments{
\item{id}{shiny input/output id}

\item{analysis}{glmm model analysis}
}
\description{
Create server logic to create interpretations for generalized linear mixed-effects models
}
\details{
This function is placed in an R shiny
server function with a specified id and model analysis.
Creates a dropdown of variables and interpretations for generalized linear mixed-effects models
}
