% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleScrollPlot.R
\name{serverScrollPlot}
\alias{serverScrollPlot}
\title{ScrollPlot Server Logic}
\usage{
serverScrollPlot(id, ggplot_list)
}
\arguments{
\item{id}{shiny input/output id}

\item{ggplot_list}{a list of ggplot objects}
}
\description{
Create server logic for scrolling ggplot outputs
}
\details{
Paired with uiScrollPlot through the id argument.
This function is placed in an R shiny
server function with a specified id and list of ggplot objects.
When rendered in combination with uiScrollPlot, creates a dynamic
ggplot object with actionButtons that scroll through the list.

Behaves dynamically depending on the length of the argument ggplot_list.
\itemize{
\item Lenght 0 will not show any plots/buttons
\item Length 1 will show only the plot
\item Length >1 will show all plots and buttons
}
}
