% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleStudyInfo.R
\name{serverStudyInfo}
\alias{serverStudyInfo}
\title{Study Information Server Logic}
\usage{
serverStudyInfo(id, type)
}
\arguments{
\item{id}{shiny input/output id}

\item{type}{A type of study}
}
\description{
Create server logic for study design information
}
\details{
Study information is the same as those supplied in the dendrogram
}
