% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleUploadInfo.R
\name{serverUploadInfo}
\alias{serverUploadInfo}
\title{moduleUploadInfo Server Logic}
\usage{
serverUploadInfo(id, data, cat_vars, cont_vars, ci_level)
}
\arguments{
\item{id}{Shiny input/output id}

\item{data}{Dataframe}

\item{cat_vars}{Categorical variables}

\item{cont_vars}{Continuous variables}

\item{ci_level}{Credible interval level}
}
\description{
Create server logic for generating list of information and warnings associated
with data upload
}
\details{
Paired with uiUploadInfo through the id argument.
This function is placed in an R shiny
server function with a user-specified parameters.

Generates a list of warnings and information about the data upload.
This includes:
\itemize{
\item Number of observations
\item Number of variables
\item Insufficient included observations
\item Variables specified as both categorical and continuous
\item Invalid CI levels
}
}
\seealso{
\code{\link{uiUploadInfo}}
}
