% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test}
\alias{t_test}
\title{Bayesian T Test}
\usage{
t_test(
  data,
  outcome_var,
  group_var,
  two_sample = FALSE,
  two_sample_group_vals,
  one_sample_group_val = NULL,
  ci_level = NULL
)
}
\arguments{
\item{data}{dataframe/tibble}

\item{outcome_var}{column in data referring to outcome of interest}

\item{group_var}{column in data that contains group labels}

\item{two_sample}{T/F indicating two sample comparison}

\item{two_sample_group_vals}{values of the group variable corresponding to two populations for comparison}

\item{one_sample_group_val}{Group value for one sample test}

\item{ci_level}{Credible interval level}
}
\description{
Bayesian T Test
}
\details{
Conducts bayesian t-test using bayesics::t_test_b function. There are 3 main setups for
testing:

Two Sample - Conducts t-test on outcome variable, grouped by the values in two_sample_group_vals
One Sample - Conducts t-test on outcome variable for the entire data column
One Sample (by group) - Conducts t-test on outcome variable but only for the
group specified by one_sample_group_val
}
\examples{
# t_test(data = iris, outcome_var = "Sepal.Length", group_var = "Species", two_sample = TRUE, one_sample_group_val = c("setosa", "virginica")
# t_test(data = iris, outcome_var = "Sepal.Length", group_var = "Species", two_sample = FALSE, one_sample_group_val = "virginica")
# t_test(data = iris, outcome_var = "Sepal.Length", group_var = "Species", two_sample = FALSE)
}
