% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleCatScrollPlot.R
\name{uiCatScrollPlot}
\alias{uiCatScrollPlot}
\title{CatScrollPlot UI Logic}
\usage{
uiCatScrollPlot(id, label = "catscrollplot")
}
\arguments{
\item{id}{shiny input/output id}
}
\description{
Create UI logic for scrolling plotly outputs for catinuous variables
}
\details{
Paired with serverCatScrollPlot through the id argument.
}
\seealso{
\code{\link{uiCatScrollPlot}}
}
