% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleDataTable.R
\name{uiDataTable}
\alias{uiDataTable}
\title{DataTable UI Logic}
\usage{
uiDataTable(id, label = "datatable")
}
\arguments{
\item{id}{shiny input/output id}
}
\description{
Create UI logic for creating data table
}
\details{
Paired with serverDataTable through the id argument.
This function is placed in an R shiny
ui function with a paired serverDataTable which shows
the data table associated with the data uploaded by the user.
}
