% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleFiveNum.R
\name{uiFiveNum}
\alias{uiFiveNum}
\title{ContScrollPlot UI Logic}
\usage{
uiFiveNum(id)
}
\arguments{
\item{id}{shiny input/output id}
}
\description{
Create UI logic for scrolling plotly outputs for continuous variables
}
\details{
Paired with serverContScrollPlot through the id argument.
This function is placed in an R shiny
ui function with a paired serverContScrollPlot.
Creates a dynamic
plotly object with actionButtons that scroll through the list.
}
