% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleFlowchart.R
\name{uiFlowchart}
\alias{uiFlowchart}
\title{Flowchart UI Logic}
\usage{
uiFlowchart(id, label = "flowchart")
}
\arguments{
\item{id}{shiny input/output id}
}
\description{
Create UI logic for rendering modified binary flowchart
}
\details{
Paired with serverFlowchart through the id argument.
}
