% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleScrollPlot.R
\name{uiScrollPlot}
\alias{uiScrollPlot}
\title{ScrollPlot UI Logic}
\usage{
uiScrollPlot(id, label = "scrollplot")
}
\arguments{
\item{id}{shiny input/output id}
}
\description{
Create UI logic for scrolling ggplot outputs
}
\details{
Paired with serverScrollPlot through the id argument.
This function is placed in an R shiny
ui function with a paired serverScrollPlot.
Creates a dynamic
ggplot object with actionButtons that scroll through the list.
}
