% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleStudyInfo.R
\name{uiStudyInfo}
\alias{uiStudyInfo}
\title{Study Information UI Logic}
\usage{
uiStudyInfo(id, label = "studyinfo")
}
\arguments{
\item{id}{shiny input/output id}
}
\description{
Create UI logic study information
}
\details{
Study information is the same as those supplied in the dendrogram
}
