% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleUploadInfo.R
\name{uiUploadInfo}
\alias{uiUploadInfo}
\title{moduleUploadInfo UI Logic}
\usage{
uiUploadInfo(id, label = "uploadinfo")
}
\arguments{
\item{id}{Shiny input/output id}
}
\description{
Create UI logic for generating list of information and warnings associated
with data upload
}
\details{
Paired with serverUploadInfo through the id argument.

\code{\link{serverUploadInfo}}
}
