% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DANA.R
\name{assessNormalization}
\alias{assessNormalization}
\title{Data-driven miRNA sequencing Normalization Assessment}
\usage{
assessNormalization(raw, normalized, negControls, posControls, clusters)
}
\arguments{
\item{raw}{Raw read count matrix (rows = genes, cols = samples).
The rows and columns of the count matrix must be named,
where \code{rownames(raw)} are the marker names
and \code{colnames(raw)} are the sample names.}

\item{normalized}{Named list of normalized count matrices.
Each matrix holds the normalized read count matrix corresponding to a
normalization method under study.
Each list member must be named (e.g. after the used normalization).
Each matrix in \code{normalized} must be named where
the row names are the marker names
and the column names are the sample names.
A list of normalized counts can be generated using
the \code{\link{applyNormalization}} function.}

\item{negControls}{Vector of negative control markers as generated by
the function \code{\link{defineControls}}.}

\item{posControls}{Vector of positive control markers as generated by
the function \code{\link{defineControls}}.}

\item{clusters}{Named Vector of clusters. Associates each miRNA
in \code{raw} to a polycistronic cluster. Usually generated using
the function \code{\link{defineClusters}}.}
}
\value{
DANA Assessment metrics for the provided normalized counts (for each
  normalized count matrix).
  DANA computes two assessment metrics:
\describe{
  \item{cc}{\code{cc} measures the preservation of biological signals
  before versus after normalization.
  A high value indicates a high preservation of biological signals
  (\code{cc} <= 1).
  In particular, \code{cc} is the concordance correlation coefficient of the
  within-cluster partial correlation among positive controls before and
  after normalization.}
  \item{mscr}{\code{mscr} measures the relative reduction of handling before
  versus after normalization. A high \code{mscr} indicates higher
  removal of handling effects.
  In particular, \code{mscr} is the mean-squared correlation reduction in
  negative controls before and after normalization.}
}
When selecting a normalization method for the \code{raw} data, one should
aim for the best possible trade-off of hight cc and high mscr.
}
\description{
Data-driven miRNA sequencing Normalization Assessment
}
